refactoring sections of form, adding tests
