+++ app/src/Pages/Portal/Contributions/ContributionReady/ContributionReady.js
@@ -2,14 +2,8 @@ import React, { Component } from "react";
 import PageHoc from "../../../../components/PageHoc/PageHoc";
 import AddContributionForm from "../../../../components/Forms/AddContribution/index";
 import { connect } from "react-redux";
-import { login } from "../../../../state/ducks/auth";
 
 class AddContribution extends Component {
-  componentWillUpdate(newprops) {
-    if (!(typeof newprops.state.me == "undefined" || !newprops.state.me)) {
-      this.props.history.push("/dashboard");
-    }
-  }
   render() {
     return (
       <PageHoc>
@@ -18,15 +12,4 @@ class AddContribution extends Component {
     );
   }
 }
-
-export default connect(
-  state => {
-    return { state: state.auth };
-  },
-  dispatch => {
-    return {
-      login: (email, password) => dispatch(login(email, password)),
-      dispatch
-    };
-  }
-)(AddContribution);
+export default connect()(AddContribution);
