+++ app/src/components/Forms/Utils/FormsUtils.test.js
@@ -1,0 +1,11 @@
+import { checkNoEmptyString } from './FormsUtils';
+
+describe("checkNoEmptyString()", () => {
+    it("should return true", () => {
+        expect(checkNoEmptyString("a", "b", "c")).toBeTruthy()
+    })
+
+    it("should return false", () => {
+        expect(checkNoEmptyString("a", "", "c")).toBeFalsy()
+    })
+})
